feat: updated scripts

+++ src/index.ts
@@ -10,20 +10,26 @@ const ESCAPE_M2 = (text: string) =>
 
 const { BOT_TOKEN, GROUP_ID, ADMIN_ID } = process.env;
 if (!BOT_TOKEN || !GROUP_ID || !ADMIN_ID) {
-  throw new Error("ENV variables aren't configured!");
+  throw new Error("ENV variables aren't configured");
 }
 
 const bot = new Telegraf(BOT_TOKEN);
-let lastMessageId: number | null = null;
 const refreshProxy = async (): Promise<void> => {
   try {
     const sh = `
       DOMAIN="ya.ru"
       PORT="9443"
 
-      IP=$(curl -s4 https://api.ipify.org)
       SECRET=$(docker run --rm nineseconds/mtg:2 generate-secret --hex $DOMAIN)
-      
+      IP=$(curl -s4 -m 3 https://api.ipify.org || \
+           curl -s4 -m 3 https://ifconfig.me || \
+           curl -s4 -m 3 https://checkip.amazonaws.com || \
+           echo "0.0.0.0")
+      if [ "$IP" = "0.0.0.0" ]; then
+        echo "Couldn't retrieve IP from providers."
+        exit 1
+      fi
+
       docker rm -f mtproto-proxy >/dev/null 2>&1 || true
 
       docker run -d \
@@ -37,11 +43,10 @@ const refreshProxy = async (): Promise<void> => {
     `;
 
     const { stdout } = await EXEC_PROMISE(sh);
-
     const lines = stdout.trim().split("\n");
-    const proxy = lines[lines.length - 1]?.trim();
-    if (!proxy?.startsWith("tg://")) {
-      console.log("Proxy generation failed:", stdout);
+    const proxy = lines.find(line => line.trim().startsWith("tg://"));
+    if (!proxy) {
+      console.log("Proxy is failed: ", stdout);
       return;
     }
 
@@ -63,41 +68,48 @@ const refreshProxy = async (): Promise<void> => {
   }
 };
 
-// 0 * * * *
-cron.schedule("*/30 * * * *", async () => {
-  await refreshProxy();
-});
-
 bot.command("refresh", async (ctx) => {
   if (ctx.from.id.toString() !== ADMIN_ID) {
-    await ctx.reply("Not enough rights.", {
+    await ctx.reply("Not enough rights!", {
       reply_parameters: { message_id: ctx.message.message_id },
     });
     return;
   }
 
-  const msg = await ctx.reply("Refreshing MTProto...", {
+  const msg = await ctx.reply("Refreshing MTProto... please wait ~10s =)", {
     reply_parameters: { message_id: ctx.message.message_id },
   });
 
-  try {
-    await refreshProxy();
-    await ctx.deleteMessage(msg.message_id);
-  } catch (err) {
-    console.log(err);
+  refreshProxy()
+    .then(async () => {
+      try {
+        await ctx.deleteMessage(msg.message_id);
+      } catch (err) {}
+    })
+    .catch(async (err) => {
+      console.error(err);
 
-    await ctx.reply("Server is not responding...");
-  }
+      await ctx.reply("Server is not responding...");
+    });
 });
 
-bot
-  .launch()
-  .then(() => {
-    console.log("Bot now lives in NL!");
+(async () => {
+  try {
+    bot.launch().catch((err) => {
+      console.log(err)
+
+      process.exit(1);
+    });
 
     refreshProxy();
-  })
-  .catch((err) => console.error(err));
+
+    cron.schedule("*/30 * * * *", async () => {
+      refreshProxy().catch((err) => console.error(err));
+    });
+  } catch (err) {
+    console.error(err);
+  }
+})();
 
 process.once("SIGINT", () => bot.stop("SIGINT"));
 process.once("SIGTERM", () => bot.stop("SIGTERM"));
